+++ CHANGELOG.md
@@ -16,9 +16,14 @@
 
 ### Bugs fixed:
 * Invoice field mapping profile: Vendor name lookup is hidden (UIDATIMP-1290)
+* Order and Order line field mapping: View screen updates (UIDATIMP-1304)
+* Order and Order line field mapping: Create/Edit screen updates (UIDATIMP-1305)
+
+## [5.3.6](https://github.com/folio-org/ui-data-import/tree/v5.3.6) (2022-11-24)
+
+### Bugs fixed:
 * Missing associated profiles on edit screen on Action/Field mapping profiles (UIDATIMP-1296)
 * Data import settings: Shortcuts for creating, editing and duplicating don't work (UIDATIMP-1300)
-* Order and Order line field mapping: Create/Edit screen updates (UIDATIMP-1305)
 * Error in module after creating new match profile (UIDATIMP-1319)
 
 ## [5.3.5](https://github.com/folio-org/ui-data-import/tree/v5.3.5) (2022-11-17)